+++ grid/hex_grid/axial_point_fn.h
@@ -14,7 +14,7 @@ namespace grid::hex
 	 * @param[in] range Enclosing distance toRange of neighbors
 	 * @return The neighboring points within range of the origin point. 
 	 */
-	std::vector<axial_point> get_neighbors(axial_point const & origin, std::int32_t range);
+	std::vector<axial_point> get_neighbors(axial_point const & origin, std::int64_t range);
 
 	/**
 	 * Get the distance between 2 points.
@@ -23,5 +23,5 @@ namespace grid::hex
 	 * @param[in] b Second point.
 	 * @return The distance between the points.
 	 */
-	std::uint32_t getDistance(axial_point const & a, axial_point const & b);
+	std::uint64_t getDistance(axial_point const & a, axial_point const & b);
 }
